+++ actionsvc/src/main/java/uk/gov/ons/ctp/response/action/service/impl/FeedbackServiceImpl.java
@@ -49,7 +49,7 @@ public class FeedbackServiceImpl implements FeedbackService {
     // TODO won't need this check when instruction xsd has responseRequired bool
     // here now so that for 2016 we could use CSV ingest for all handlers if necessary
     if (actionId.compareTo(CSV_GENERATED_ID_BOUNDARY) == -1) {
-      Action action = actionRepo.getOne(actionId.intValue());
+      Action action = actionRepo.getOne(actionId);
       ActionDTO.ActionEvent outcomeEvent = ActionDTO.ActionEvent.valueOf(feedback.getOutcome().name());
 
       if (outcomeEvent != null) {
